Create benchmarks for DisplayListBuilder
